Change main.py to new objective
This commit changes the main file to reflect my new vision for this project. I'm essentially tossing out everything before this in favor of this new approach, where the song is generated first and then played. This allows for a much more consistent song and allows me to do other things with the music.